Move cVEP common functions to datasets.utils - Closes #564
